Broke handlers up into feeds and sync modules.  Created parsers for Order and Contract objects.
